Refactored base problem class AbstractProblem
